requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_community
langchain_groq
pinecone
pinecone-text
sentence-transformers
huggingface-hub
langchain-huggingface
transformers
torch
streamlit
speechrecognition 
gtts
pygame
flask

